requirements.txt
# Core dependencies
tensorflow>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-learn>=1.0.0

# EDA specific
jupyterlab>=3.0.0
plotly>=5.0.0
seaborn>=0.11.0
statsmodels>=0.13.0
yellowbrick>=1.4.0
ipywidgets>=7.6.0
tqdm>=4.62.0

# Data augmentation
albumentations>=1.1.0

# Face detection
mtcnn>=0.1.1

# Configuration
pyyaml>=5.4.0

# API (optional)
fastapi>=0.70.0
uvicorn>=0.15.0

# Utilities
matplotlib>=3.4.0
jinja2>=3.0.0

# Development
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0







